backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
channels>=4.0.0
daphne>=4.0.0
redis>=5.0.0
gunicorn>=21.2.0
requests>=2.31.0
